+++ Formula/netductor.rb
@@ -1,30 +1,30 @@
 class Netductor < Formula
-  desc "Netductor"
+  desc "Netductor control plane CLI / TUI"
   homepage "https://github.com/PavelNeyman/netductor"
-  version "0.8.93"
+  version "0.8.94"
   license "MIT"
   on_macos do
     on_arm do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.93/netductor-darwin-arm64"
-      sha256 "e174db64f0cc42dd3b270864f5be523ee95bd5d69013d04de4db177bb98b8c1a"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.94/netductor-darwin-arm64"
+      sha256 "73ab9d39701158550fa1100a56991c47e780d61a4e46bd56290fff80dac212ee"
     end
     on_intel do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.93/netductor-darwin-amd64"
-      sha256 "35edfce94098d39669d029da0fa626d6651a812f2c25600bf48784a1b5c90387"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.94/netductor-darwin-amd64"
+      sha256 "46d526a0fc1dcfa65490b8294c51af3b737707e01734ea1ae78598a49c7ff54d"
     end
   end
   on_linux do
     on_intel do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.93/netductor-linux-amd64"
-      sha256 "9b44a49670714ed6d1b0b54219b7fe51df3f6245f6f4fcd6a83c10795239f6bc"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.94/netductor-linux-amd64"
+      sha256 "70e01301549fd3211f36da1bf89b030f86ca9da76e6511999d6762b83e799f23"
     end
     on_arm do
-      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.93/netductor-linux-arm64"
-      sha256 "c4a389f356831744cdac03112ef868d77e64484fe555574ae8a75d4c7a96118d"
+      url "https://github.com/PavelNeyman/netductor/releases/download/v0.8.94/netductor-linux-arm64"
+      sha256 "a2735705a6d65b19416329a66ad75e4fbf7f413a18d22dbc02dd871a2f2623d1"
     end
   end
   def install
-    bin.install Dir["netductor*"].first => "netductor"
+    bin.install Dir["netductor-*"].first => "netductor"
   end
   test do
     assert_match version.to_s, shell_output("#{bin}/netductor version 2>&1")